Persistencia entidad curso

+++ src/main/java/com/iesmb/gestionalumnos/repository/CursoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.iesmb.gestionalumnos.entity.Curso;
 
-public interface CursoRepository extends JpaRepository <Curso, Long>{
+public interface CursoRepository extends JpaRepository <Curso, Integer>{
 
 }
